refactor(network): replicate to generals using patch method
refactor(network): move NetPacketStructs to Core
