+++ Core/GameEngine/Source/GameNetwork/NetCommandMsg.cpp
@@ -1271,4 +1271,3 @@ NetTimeOutGameStartCommandMsg::~NetTimeOutGameStartCommandMsg() {
 size_t NetTimeOutGameStartCommandMsg::getPackedByteCount() const {
 	return sizeof(NetPacketTimeOutGameStartMessage);
 }
-
