+++ Generals/Code/GameEngine/CMakeLists.txt
@@ -502,6 +502,7 @@ set(GAMEENGINE_SRC
 #    Include/GameNetwork/NetCommandRef.h
 #    Include/GameNetwork/NetCommandWrapperList.h
 #    Include/GameNetwork/NetPacket.h
+#    Include/GameNetwork/NetPacketStructs.h
 #    Include/GameNetwork/NetworkDefs.h
 #    Include/GameNetwork/NetworkInterface.h
 #    Include/GameNetwork/networkutil.h
